+++ server/services/educationSupportPlanService.ts
@@ -1,11 +1,8 @@
-import type { PlanCreationScheduleResponse, PlanCreationSchedulesResponse } from 'supportAdditionalNeedsApiClient'
-import type { EducationSupportPlanDto, PlanCreationScheduleDto, RefuseEducationSupportPlanDto } from 'dto'
+import type { EducationSupportPlanDto } from 'dto'
 import { SupportAdditionalNeedsApiClient } from '../data'
 import toCreateEducationSupportPlanRequest from '../data/mappers/createEducationSupportPlanRequestMapper'
-import logger from '../../logger'
 import toEducationSupportPlanDto from '../data/mappers/educationSupportPlanDtoMapper'
-import toPlanCreationScheduleDto from '../data/mappers/planCreationScheduleDtoMapper'
-import toUpdatePlanCreationStatusRequest from '../data/mappers/updatePlanCreationStatusRequestMapper'
+import logger from '../../logger'
 
 export default class EducationSupportPlanService {
   constructor(private readonly supportAdditionalNeedsApiClient: SupportAdditionalNeedsApiClient) {}
@@ -24,70 +21,5 @@ export default class EducationSupportPlanService {
       logger.error('Error creating education support plan', e)
       throw e
     }
-  }
-
-  /**
-   * Returns the prisoner's current (most recent) Education Support Plan Creation Schedule (PlanCreationScheduleDto),
-   * or null if the prisoner does not have a Plan Creation Schedule setup.
-   *
-   * If the most recent Plan Creation Schedule is COMPLETED or EXEMPT_ it will be returned and the calling code should
-   * determine what to do in this case.
-   */
-  async getCurrentEducationSupportPlanCreationSchedule(
-    prisonNumber: string,
-    username: string,
-  ): Promise<PlanCreationScheduleDto> {
-    try {
-      const includeAllHistory = false
-      const apiResponse = await this.supportAdditionalNeedsApiClient.getEducationSupportPlanCreationSchedules(
-        prisonNumber,
-        username,
-        includeAllHistory,
-      )
-
-      const planCreationSchedules: Array<PlanCreationScheduleResponse> = apiResponse?.planCreationSchedules || []
-      if (planCreationSchedules.length === 0) {
-        return null
-      }
-
-      return this.currentPlanCreationSchedule(prisonNumber, apiResponse)
-    } catch (e) {
-      logger.error('Error retrieving education support plan creation schedule', e)
-      throw e
-    }
-  }
-
-  async updateEducationSupportPlanCreationScheduleAsRefused(
-    username: string,
-    dto: RefuseEducationSupportPlanDto,
-  ): Promise<PlanCreationScheduleDto> {
-    try {
-      const updatePlanCreationStatusRequest = toUpdatePlanCreationStatusRequest(dto)
-      const apiResponse = await this.supportAdditionalNeedsApiClient.updateEducationSupportPlanCreationScheduleStatus(
-        dto.prisonNumber,
-        username,
-        updatePlanCreationStatusRequest,
-      )
-
-      return this.currentPlanCreationSchedule(dto.prisonNumber, apiResponse)
-    } catch (e) {
-      logger.error('Error updating education support plan creation schedule as Refused', e)
-      throw e
-    }
-  }
-
-  private currentPlanCreationSchedule = (
-    prisonNumber: string,
-    apiResponse: PlanCreationSchedulesResponse,
-  ): PlanCreationScheduleDto => {
-    const { planCreationSchedules }: { planCreationSchedules: Array<PlanCreationScheduleResponse> } = apiResponse
-    const currentSchedule =
-      planCreationSchedules.length === 1
-        ? planCreationSchedules.at(0)
-        : planCreationSchedules //
-            .toSorted((left, right) => left.version - right.version)
-            .toReversed()
-            .at(0)
-    return toPlanCreationScheduleDto(prisonNumber, currentSchedule)
   }
 }
